requirements.txt
streamlit==1.47.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.0
joblib==1.4.2
tqdm==4.66.4
requests==2.31.0
python-dotenv==1.0.1
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.3
notebook
ipykernel